Create travel_time_distance_google.ipynb
Notebook processes 65x65 and PC4 data to create weighted centroids. It requests the travel time and distance between the weighted centroids in the Rotterdam urban area using the Google Maps API.

A pickle file with the data is saved.

Warning: Be careful with the Google API. API calls aren't free above 40.000 elements monthly ($200).