Initial car scheduling with UXsim
Initial car scheduling with UXsim. It draws a random node close by (in its mrdh65 area), if any is available. If not the trip might not be relevant at all, but that needs to be researched further.

This needs a special version of UXsim: https://github.com/EwoutH/UXsim/tree/patch-5