model: Value of time per mode, av_factor
Adds value of times per mode, to make the mode choice model more realistic. Because of that transit VoT is quite low, it helps to get transit in the right ball park.

It also adds a av_vot_factor, which is by default one, meaning that the value of time for av is the same as for cars. However, it can be modified (in experiments, to see at which point people are moving other modalities to the self-driving car.

With a vot factor 1, the av isn't attractive at all.

Mode shares: ['car: 16.09%', 'bike: 75.08%', 'transit: 8.69%', 'av: 0.14%']

https://www.kimnet.nl/binaries/kimnet/documenten/publicaties/2023/12/04/nieuwe-waarderingskengetallen-voor-reistijd-betrouwbaarheid-en-comfort/Significance_Value+of+Travel+Time+in+the+Netherlands+2022_final+technical+report.pdf