Upload pc4 travel time and distance matrices
Upload pc4 travel time and distance matrices that were collected for all 125 pc4 locations in the city area. This was extensively tested and validated.

These were very expensive. 5 USD per 1000 elements. The full 125x125 matrix is 78.125 USD, so for both modes it is 156.25 USD. Don't do again.

The data was collected for for 2024-09-017 08:00:00 (a Thursday) were used.