Add option for AV density
The AV density increases or decreases the number of cars generated in the simulation. Basically, with an AV density of 1.2, 20% additional trips are generated, while with 0.5 only half are. These can be interpreted as car sharing (two people taking one car) or as the fact that AVs need to drive between customers, to charging stations, etc.