Implement edge speed reduction policy
Implement an optional policy that reduces the speed on a fraction (`policy_speed_reduction`) of the roads in a certain area (`policy_area`). It reduces speed with 20 km/h, with a minimum of 20 km/h. It also increases density on these edges with 0.02, considering slower traffic can have higher densities.