CRITICAL: Fix network speed limit bug
Critical bugfix in the assignments of speed limits in the network. OSMnx saves speed limits as `maxspeed` and `speed_kph`. `speed` wasn't available, so all roads got an edge speed of 30.

This commit fixes that. Car travel times are expected to be significantly shorter, and thus perceived car and av costs and mode share to be significantly higher.

If multiple speeds are present because the links got merged, it takes the median value.